% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ls_fun_calls}
\alias{ls_fun_calls}
\title{Lista ll function calls in a source package}
\usage{
ls_fun_calls(path = ".")
}
\arguments{
\item{path}{A path to a root directory of a package
List all named function delcarations}
}
\description{
Lista ll function calls in a source package
}
