% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ls_fun_decs}
\alias{ls_fun_decs}
\title{What functions are delared?}
\usage{
ls_fun_decs(path = ".")
}
\arguments{
\item{path}{A path to a root directory of a package
List all named function delcarations}
}
\description{
List all named function declarations in a package.
}
\details{
Returns a tibble with:

* the names of all named function declarations made under `R/` (column
  `text`).
* where exactly they are defined (column `source`).
* on what line they start (column `line1`).
* and many lines long they are (column `n_lines`).
}
