% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsesum-package.R
\docType{package}
\name{parsesum-package}
\alias{parsesum}
\alias{parsesum-package}
\title{parsesum: Analyse source code through parsing}
\description{
This package builds on top of the R base parser to derive
different figures of interest of a project a parser is suited to answer.
How many functions are delcared in this project? What's their name? How
How many lines of code per function / in total? Which functions are called
most often? And so on...
}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
\keyword{internal}
